signaling-server/requirements.txt
aiortc==1.4.0
websockets==11.0
PyYAML
